indent Mock/* and type changing in Position.h

+++ src/type/Position.h
@@ -4,7 +4,7 @@
 class Position
 {
  public:
-  int x,y,z,rx,ry,rz;
+  double x,y,z,rx,ry,rz;
 };
 
 #endif
